initial create and added models to negotiation module.
