+++ app/negotiation/models.py
@@ -1,1 +1,45 @@
 from app import db
+from sqlalchemy import Column, Date, DateTime, Float, ForeignKey, Integer, String, TIMESTAMP, text
+from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy.orm import relationship
+from sqlalchemy.ext.declarative import declarative_base
+
+Base = declarative_base()
+metadata = Base.metadata
+
+class Negotiation(Base):
+    __tablename__ = 'negotiation'
+
+    negotiation_id = Column(INTEGER, primary_key=True, unique=True)
+    vehical_id = Column(ForeignKey('vehical.vehical_id'), nullable=False, index=True)
+    customer_id = Column(ForeignKey('customer.customer_id'), nullable=False, index=True)
+    negotiation_status = Column(String(45), nullable=False, server_default=text("'Active'"))
+    start_date = Column(DateTime, server_default=text("CURRENT_TIMESTAMP"))
+    end_date = Column(DateTime)
+
+    customer = relationship('Customer')
+    vehical = relationship('Vehical')
+
+class Offer(Base):
+    __tablename__ = 'offer'
+
+    offer_id = Column(INTEGER, primary_key=True, unique=True)
+    negotiation_id = Column(ForeignKey('negotiation.negotiation_id'), nullable=False, index=True)
+    offer_price = Column(Integer, nullable=False)
+    offer_date = Column(DateTime, server_default=text("CURRENT_TIMESTAMP"))
+    offer_status = Column(String(45))
+
+    negotiation = relationship('Negotiation')
+
+
+
+class CounterOffer(Base):
+    __tablename__ = 'counter_offer'
+
+    counter_offer_id = Column(INTEGER, primary_key=True, unique=True)
+    offer_id = Column(ForeignKey('offer.offer_id'), nullable=False, unique=True)
+    counter_price = Column(Integer, nullable=False)
+    counter_date = Column(DateTime, server_default=text("CURRENT_TIMESTAMP"))
+    counter_status = Column(String(45))
+
+    offer = relationship('Offer')
